+++ lib/rewrites.php
@@ -2,14 +2,13 @@
 /**
  * URL rewriting
  *
- * Rewrites currently do not happen for child themes (or network installs)
- * @todo https://github.com/retlehs/roots/issues/461
+ * Rewrites do not happen for multisite installations or child themes
  *
  * Rewrite:
- *   /wp-content/themes/themename/css/ to /css/
- *   /wp-content/themes/themename/js/  to /js/
- *   /wp-content/themes/themename/img/ to /img/
- *   /wp-content/plugins/              to /plugins/
+ *   /wp-content/themes/themename/assets/css/ to /assets/css/
+ *   /wp-content/themes/themename/assets/js/  to /assets/js/
+ *   /wp-content/themes/themename/assets/img/ to /assets/img/
+ *   /wp-content/plugins/                     to /plugins/
  *
  * If you aren't using Apache, alternate configuration settings can be found in the docs.
  *
@@ -18,9 +17,7 @@
 function roots_add_rewrites($content) {
   global $wp_rewrite;
   $roots_new_non_wp_rules = array(
-    'assets/css/(.*)'      => THEME_PATH . '/assets/css/$1',
-    'assets/js/(.*)'       => THEME_PATH . '/assets/js/$1',
-    'assets/img/(.*)'      => THEME_PATH . '/assets/img/$1',
+    'assets/(.*)'          => THEME_PATH . '/assets/$1',
     'plugins/(.*)'         => RELATIVE_PLUGIN_PATH . '/$1'
   );
   $wp_rewrite->non_wp_rules = array_merge($wp_rewrite->non_wp_rules, $roots_new_non_wp_rules);
@@ -28,14 +25,14 @@ function roots_add_rewrites($content) {
 }
 
 function roots_clean_urls($content) {
-  if (strpos($content, FULL_RELATIVE_PLUGIN_PATH) === 0) {
-    return str_replace(FULL_RELATIVE_PLUGIN_PATH, WP_BASE . '/plugins', $content);
+  if (strpos($content, RELATIVE_PLUGIN_PATH) > 0) {
+    return str_replace('/' . RELATIVE_PLUGIN_PATH,  '/plugins', $content);
   } else {
     return str_replace('/' . THEME_PATH, '', $content);
   }
 }
 
-if (!is_multisite() && !is_child_theme() && get_option('permalink_structure')) {
+if (!is_multisite() && !is_child_theme()) {
   if (current_theme_supports('rewrites')) {
     add_action('generate_rewrite_rules', 'roots_add_rewrites');
   }