+++ lib/sidebar.php
@@ -30,7 +30,7 @@ class Roots_Sidebar {
 
   private function check_conditional_tag($conditional_tag) {
     if (is_array($conditional_tag)) {
-      return call_user_func_array($conditional_tag[0], $conditional_tag[1]);
+      return $conditional_tag[0]($conditional_tag[1]);
     } else {
       return $conditional_tag();
     }